Update bulk ediitng
